Editing files organisation
* Editing JavaScript files place
Because there is only one JS file and one HTML file per folder, it is
most appropriate to place them in a single and same folder. This way, we
avoid having to dive too deep into the folder hierarchy.

* Deleting Roman Numbers folder from repository
Because Roman Arabic Numeral Converter project covers Roman Numbers, it
is redundant and should be deleted.